+++ src/main/java/model/piece/goongsungpiece/Sa.java
@@ -1,9 +1,9 @@
-package model.piece.PiecesInGoongsung;
+package model.piece.goongsungpiece;
 
 import model.Team;
 import model.piece.PieceName;
 
-public class Sa extends PiecesInGoongsung {
+public class Sa extends GoongsungPiece {
 
     public Sa(Team team) {
         super(team);
