+++ src/test/java/model/piece/JanggiBoardTest.java
@@ -1,4 +1,4 @@
-package model;
+package model.piece;
 
 import static model.janggiboard.JanggiBoardSetUp.INNER_ELEPHANT;
 import static org.assertj.core.api.Assertions.assertThat;
@@ -10,7 +10,7 @@ import org.junit.jupiter.api.Test;
 public class JanggiBoardTest {
     @Test
     @DisplayName("장기말 세팅 테스트 - 초기 장기말 갯수 확인")
-   void  test1() {
+    void test1() {
         JanggiBoard janggiBoard = new JanggiBoard(INNER_ELEPHANT);
 
         assertThat(janggiBoard.countPiece()).isEqualTo(32);
